изменил интерфейс уроков, убрал класс tasks сделал TreeNode подклассом Tree перенес методы вставки и удаления в класс TreeNode, а методы работы с деревом в классе Tree
