+++ AlgoritmQuests/ILessons.cs
@@ -8,8 +8,10 @@ namespace AlgoritmQuests
 {
     interface ILessons
     {
-        
+        public string NameTask { get; set; }
+        public string Description { get; set; }
         void StartTask();
+        void ShowTask();
         void TaskLogic();
     }
 }
