+++ AlgoritmQuests/Program.cs
@@ -1,61 +1,42 @@
 ﻿using System;
-
+using System.Collections.Generic;
 
 namespace AlgoritmQuests
 {
     class Program
     {
-        static void Lessons()
+        static string Lessons()
         {
-            Tasks lessons = new Tasks(1);
-            //изменение шапки для отбражения уроков
-            lessons = lessons.ChangeTask(0, "Номер урока", "Краткое содержание");
-            //Урок 1
-            lessons = lessons.AddTask("Урок №1", "Блок-схемы, асимптотическая сложность, рекурсия.");
-            //Урок 2
-            lessons = lessons.AddTask("Урок №2", "Массив, список, поиск.");
-            //Урок 3
-            lessons = lessons.AddTask("Урок №3", "Класс, структура и дистанция.");
-            //Урок 4
-            lessons = lessons.AddTask("Урок №4", "Деревья, хэш-таблицы.");
-            int numLesson = lessons.ShowTask(); //вывожу на экран уроки и получаю номер выбранного урока
-
-            switch (numLesson)
-            {
-                case 1:
-                    {
-
-                        Task1 lesson1 = new Task1();
-                        lesson1.ShowTask();
-                        break;
-                    }
-                case 2:
-                    {
-
-                        Task2 lesson2 = new Task2();
-                        lesson2.ShowTask();
-                        break;
-                    }
-                case 3:
-                    {
-
-                        Task3 lesson3 = new Task3();
-                        lesson3.StartTask();
-                        break;
-                    }
-                case 4:
-                    {
-
-                        Task4 lesson4 = new Task4();
-                        lesson4.StartTask();
-                        break;
-                    }
-            }
-
+            List<ILessons> lessons = new List<ILessons>();
+            lessons.Add(new Task1 { NameTask = "Урок №1", Description = "Блок-схемы, асимптотическая сложность, рекурсия." });
+            lessons.Add(new Task2 { NameTask = "Урок №2", Description = "Массив, список, поиск." });
+            lessons.Add(new Task3 { NameTask = "Урок №3", Description = "Класс, структура и дистанция." });
+            lessons.Add(new Task4 { NameTask = "Урок №4", Description = "Деревья, хэш-таблицы." });
+            lessons.Add(new Task5 { NameTask = "Урок №5", Description = "Стек, очередь, словарь и коллекции в C#" });
+            for (int i = 0; i < lessons.Count; i++) Console.WriteLine(lessons[i].NameTask + "\n" + lessons[i].Description + "\n"); 
+            Console.WriteLine("Введите номер интересующего задания и нажмите Enter: ");
+            string input = Console.ReadLine();
+            int N = 0;
+            bool successChange = int.TryParse(input,out N);
+            if (successChange & (N > 0) & (N<=lessons.Count)) lessons[N-1].StartTask();
+            return input;
         }
         static void Main(string[] args)
         {
-            Lessons();
+            string input = ""; 
+            do
+            {
+                Console.WriteLine("Для завершения введите exit \n");
+                input = Lessons();
+                if (input != "exit")
+                {
+                    Console.WriteLine("Для продолжение нажмите Enter \n");
+                    Console.ReadLine();
+                    Console.Clear();
+                }
+            }
+            while (input != "exit") ;
+
         }
     }
 }
